pyproject.toml
[project]
name = "nectarapp"
version = "0.1.0"
description = "Property management API"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.103.0",
    "uvicorn[standard]>=0.22.0",
    "sqlalchemy>=2.0.0",
    "databases[postgresql]>=0.9.0",
    "passlib[bcrypt]",
    "python-jose[cryptography]",
    "python-dotenv",
    "alembic",
    "asyncpg",
    "psycopg2-binary",
    "pydantic-settings>=2.0.0",
    "python-multipart"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"